// src/testimoni/dto/create-testimoni.dto.ts
export class CreateTestimoniDto {
    id?: number;
    profilePicturePath?: string;
    name: string;
    email?: string;
    position?: string;
    company?: string;
    website?: string;
    testimonialText: string;
    rating?: number;
    type?: string;
    source?: string;
    path?: string;
    projectId?: number;
    formId?: number;
} 